fix(RHINENG-13688): Convert publish date filter to select
This PR removes the radios from the publish date filter dropdown on the CVEs table toolbar. It implements the custom Select filter that was created for the Create Report modal.

The publish date filter on the CVEs table does not include the "All" option. In order to get all publish dates, a user must remove any applied publish date filter.

The Create Report modal, on the other hand, does include the "All" option because there is no way for a user to remove a publish date filter if one is applied.