fix(RHINENG-12778): Add tooltip for locked OS versions
This PR notifies a user if a system's OS version is locked. When version
locking is enabled for a system i.e. a system's OS is tied to a release
version, that can affect the count of applicable and installable
errata\advisories.

The backend now passes a key from the `/systems` api endpoint
called `rhsm_lock` and the value is the locked version number. With this
PR, we check on the Systems table as well as the CVEs -> CVEDetails
tables. When the `rhsm_lock` parameter has a value, we include an icon
with a tooltip that tells the user "Your RHEL version is locked..." and
at which version it is locked at.